Add error handling when record does not exist

+++ src/main/java/magazynier/DAO.java
@@ -1,5 +1,6 @@
 package magazynier;
 
+import javassist.NotFoundException;
 import org.hibernate.HibernateException;
 import org.hibernate.Session;
 import org.hibernate.Transaction;
@@ -7,6 +8,7 @@ import org.hibernate.Transaction;
 import java.util.ArrayList;
 
 public class DAO {
+
     public static ArrayList readTable(String tableName) {
 
         ArrayList whl = new ArrayList();
@@ -21,19 +23,8 @@ public class DAO {
         return whl;
     }
 
-    public static void update(Object object) {
-        try (Session session = HibernateSessionFactory.openSession()) {
-            Transaction tr = session.beginTransaction();
-            session.update(object);
-            tr.commit();
-            session.close();
-        } catch (HibernateException e) {
-            System.out.println("update failed");
-            e.printStackTrace();
-        }
-    }
-
     public static void add(Object object) {
+
         try (Session session = HibernateSessionFactory.openSession()) {
             Transaction tr = session.beginTransaction();
             session.save(object);
@@ -45,15 +36,37 @@ public class DAO {
         }
     }
 
-    public static void delete(Object object) throws HibernateException {
-        Session session = HibernateSessionFactory.openSession();
-        Transaction tr = session.beginTransaction();
-        session.delete(object);
-        tr.commit();
-        session.close();
+    public static void update(Indexed object) throws NotFoundException {
+
+        if (checkIfExistsById(object.getClass(), object.getId())) {
+            try (Session session = HibernateSessionFactory.openSession()) {
+                Transaction tr = session.beginTransaction();
+                session.update(object);
+                tr.commit();
+                session.close();
+            } catch (HibernateException e) {
+                System.out.println("update failed");
+                e.printStackTrace();
+            }
+        } else {
+            throw new NotFoundException(object.getClass() + " object:" + object + " does not exist in database.");
+        }
     }
 
-    public static boolean checkIfExistsById(Class c, Integer id) {
+    public static void delete(Indexed object) throws NotFoundException {
+
+        if (checkIfExistsById(object.getClass(), object.getId())) {
+            Session session = HibernateSessionFactory.openSession();
+            Transaction tr = session.beginTransaction();
+            session.delete(object);
+            tr.commit();
+            session.close();
+        } else {
+            throw new NotFoundException(object.getClass() + " object:" + object + " does not exist in database.");
+        }
+    }
+
+    private static boolean checkIfExistsById(Class c, Integer id) {
         Object object = new Object();
         Session session = HibernateSessionFactory.openSession();
         Object obj = session.get(c, id);
